myproject/requirements.txt
dj_rest_auth==7.0.1
Django==5.2.5
django_allauth==65.10.0
django_environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
Requests==2.32.4
